move FactorModelGlobalRSI next to FactorModel
